refactor: sync create/edit hooks to use consistent object signature and fix image handling in apiCabins

+++ src/services/apiCabins.js
@@ -26,34 +26,45 @@ export async function deleteCabin(id) {
   }
 }
 
-export async function createCabin(newCabin) {
-  // 1. Create a unique name for the image to avoid name collisions
-  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
+export async function createUpdateCabin(newCabin, id) {
+  // 1. Check if we already have a URL string (means no new image upload)
+  const hasImagePath =
+    typeof newCabin.image === 'string' &&
+    newCabin.image.startsWith(supabaseUrl);
+
+  const imageName = `${Math.random()}-${newCabin.image?.name}`.replaceAll(
     '/',
     ''
   );
 
-  // 2. Create the URL path that will be stored in the database
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  // 3. Create the cabin row in the database
-  const { data, error } = await supabase
-    .from('cabins')
-    .insert([{ ...newCabin, image: imagePath }])
-    .select()
-    .single();
+  // 2. Create/edit cabin
+  let query = supabase.from('cabins');
+
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq('id', id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
-    throw new Error('Cabin could not be created');
+    throw new Error('Cabin could not be created/updated');
   }
 
-  // 4. Upload the actual image file to the bucket
+  // 3. Upload image if there is a new one
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from('cabin-images')
     .upload(imageName, newCabin.image);
 
-  // 5. Clean up: If image upload failed, delete the cabin row
+  // 4. Delete the cabin IF there was an error uploading image
   if (storageError) {
     await supabase.from('cabins').delete().eq('id', data.id);
     console.error(storageError);
